improve the efficiency of the diffBy algorithm and drop its dependency on the Vector
